refactor: publishPostForLens
